Conflicts:
	Orange/widgets/__init__.py
	Orange/widgets/data/owcolor.py
	Orange/widgets/data/owconcatenate.py
	Orange/widgets/data/owcontinuize.py
	Orange/widgets/data/owdatasampler.py
	Orange/widgets/data/owdiscretize.py
	Orange/widgets/data/oweditdomain.py
	Orange/widgets/data/owfeatureconstructor.py
	Orange/widgets/data/owfile.py
	Orange/widgets/data/owimageviewer.py
	Orange/widgets/data/owimpute.py
	Orange/widgets/data/owmergedata.py
	Orange/widgets/data/owoutliers.py
	Orange/widgets/data/owpaintdata.py
	Orange/widgets/data/owpreprocess.py
	Orange/widgets/data/owpurgedomain.py
	Orange/widgets/data/owrank.py
	Orange/widgets/data/owsave.py
	Orange/widgets/data/owselectcolumns.py
	Orange/widgets/data/owselectrows.py
	Orange/widgets/data/owtable.py

+++ Orange/widgets/utils/classdensity.py
@@ -27,7 +27,7 @@ def class_density_image(min_x, max_x, min_y, max_y, resolution, x_data, y_data, 
     y_grid_norm = (np.array(y_grid)-min_y)/(max_y-min_y)
     img = compute_density(x_grid_norm, y_grid_norm,
                           x_data_norm[sample], y_data_norm[sample], np.array(rgb_data)[sample])
-    density_img = ImageItem(img, autoLevels=False)
+    density_img = ImageItem(img.astype(np.uint8), autoLevels=False)
     density_img.setRect(QRectF(min_x-x_sz/2, min_y-y_sz/2,
                                     max_x-min_x+x_sz, max_y-min_y+y_sz))
     density_img.setZValue(-1)
